+++ src/polygons.jl
@@ -871,7 +871,7 @@ end
 """
     pathtopoly()
 
-Convert the current path to an array of polygons.
+Build a copy of the current path as an array of polygons. The current path remains active.
 
 Returns an array of polygons, corresponding to the paths and subpaths of the original path.
 """